+++ NEWS.md
@@ -3,6 +3,8 @@
   * Bug fix in plot function for univariate result.
   * Tree does not need to be ultrametric for a BM
   * Change `tol` to `tol_EM` in `estimateEM` for actual binding
+* API Change
+  * For a univariate trait, the BM re-scaling is now applied by default.
 
 # PhylogeneticEM 1.2.1
 * CRAN maintenance